requirements.txt
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
docopt==0.6.2
et-xmlfile==1.1.0
Flask==3.0.0
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
munch==4.0.0
numpy==1.26.0
openpyxl==3.1.2
packaging==23.1
pandas==2.1.1
pipreqs==0.4.13
pluggy==1.3.0
pytest==7.4.2
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
tzdata==2023.3
urllib3==2.0.6
Werkzeug==3.0.0
yarg==0.1.9
